fix code to allow visualization of the "save needed" popup .
Summary:
in a situation where you were trying to close the application without saving,
the program made it appear an empty pop-up, this prevents the user to save any unsaved data.
This is because a box was used to create popup's content and popup's buttons.

elm_object_part_content_set and elm_object_part_text_set should be used instead
to properly populate the popup

@fix

Reviewers: cedric, tasn

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D2612