bin: Correct selected item in settings
elm_toolbar and elm_naviframe order differently. First item added to a
toolbar in single select mode is selected. The last item added to a
naviframe is the first one shown. Maybe improper API usage or wonky API.
Did not come across any calls to set active/highlight/selected on toolbar
item. Thus only way to correct is ordering of creating and adding items.

Thanks to AJSlye/@NuLogicSystems for the report of the issue!