bin: Fix issue, remove files from recent that are not accessible
Initial fix for issue #30 removing files from recent files that are not
accessible. Could be modified to only remove files that do not exist,
but presently removing any file from recent that stat fails on for any
reason.

Need to switch to absolute paths vs relative. This introduces a new
issue with files being removed when the directory ecrire is running
in changes. Which will change all relative paths, and cause removal of
all in history. Thus need to switch to absolute vs relative paths.

https://github.com/Obsidian-StudiosInc/ecrire/issues/30