bin: Added function to resize inwin due to side effect from efl_part
efl_part causes inwin to be off centered as reported in issue #37 and #40
Adding evas_object_resize centers the inwin in most cases, except that
reported in #40. This function is used when the window is resized,
resize the inwin object as well. Really should not be needed, but for
some reason is due to side effects from elf_part.

[efl] efl_part function causes Inner windows to be off center
https://github.com/Obsidian-StudiosInc/ecrire/issues/37

Save dialog issues
https://github.com/Obsidian-StudiosInc/ecrire/issues/40